Finish cleanup + using negative space translations
Actually finishes Trello card b5PYUO1r: I went through and replaced every use of the unicode characters with the translation keys. This should be more future proof if the resourcepack changes, as the translation keys should stay the same.
I also moved the hud and gui characters to their own font, and made the place functions for the 2 crafting tables use them.
The crafting table textures have a 1-opacity white pixel in the bottom left, so negative spacing for the actual name of the gui can always be space.-249 regardless of the actual texture.